requirements.txt
fastapi
uvicorn
python-multipart
jinja2
aiofiles
torch
torchvision
transformers
Pillow
opencv-python
numpy